Better adaptive adjustment of the number of loop iterations
Keep doubling the number of loop iterations until the duration
of test run exceeds 0.5s and also start from 1 loop iteration
instead of 16. In the case if the memory bandwidth is extremely
low (for example, running tests with the framebuffer), it makes
the test duration reasonable. Also in the case if the memory
bandwidth is extremely high, this approach reduces the periodic
gettimeofday() calls overhead.